lesHandlingTime

+++ SmartPort/src/smartPort/Items.java
@@ -4,6 +4,6 @@ import java.util.Dictionary;
 
 public class Items {
 
-    public static Dictionary<String, Integer> items;
+    public static Dictionary<String, Float> items;
 
 }
